fix: Fix navigation link to homepage
This commit fixes a navigation link in the documentation's index.html. The link previously pointed to '#hero', which is an in-page anchor. It has been updated to './' to correctly navigate to the root of the documentation (homepage). This ensures correct navigation when the documentation is accessed from different parts of the site or directly.